+++ pybricksdev/flash.py
@@ -3,167 +3,20 @@
 
 import asyncio
 import io
-import json
 import logging
-import os
 import platform
 import struct
-import zipfile
 from collections import namedtuple
-from typing import BinaryIO, Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Tuple
 
-import semver
 from tqdm.auto import tqdm
 from tqdm.contrib.logging import logging_redirect_tqdm
 
 from .ble import BLERequestsConnection
 from .ble.lwp3.bootloader import BootloaderCommand
 from .ble.lwp3.bytecodes import HubKind
-from .compile import compile_file, save_script
-from .firmware import (
-    AnyFirmwareV1Metadata,
-    AnyFirmwareV2Metadata,
-    AnyFirmwareMetadata,
-    firmware_metadata_is_v1,
-    firmware_metadata_is_v2,
-)
-from .tools.checksum import crc32_checksum, sum_complement
 
 logger = logging.getLogger(__name__)
-
-
-async def _create_firmware_v1(
-    metadata: AnyFirmwareV1Metadata, archive: zipfile.ZipFile, name: Optional[str]
-) -> bytearray:
-    base = archive.open("firmware-base.bin").read()
-
-    if "main.py" in archive.namelist():
-        main_py = io.TextIOWrapper(archive.open("main.py"))
-
-        mpy = await compile_file(
-            save_script(main_py.read()),
-            metadata["mpy-abi-version"],
-            metadata["mpy-cross-options"],
-        )
-    else:
-        mpy = b""
-
-    # start with base firmware binary blob
-    firmware = bytearray(base)
-    # pad with 0s until user-mpy-offset
-    firmware.extend(0 for _ in range(metadata["user-mpy-offset"] - len(firmware)))
-    # append 32-bit little-endian main.mpy file size
-    firmware.extend(struct.pack("<I", len(mpy)))
-    # append main.mpy file
-    firmware.extend(mpy)
-    # pad with 0s to align to 4-byte boundary
-    firmware.extend(0 for _ in range(-len(firmware) % 4))
-
-    # Update hub name if given
-    if name:
-        if semver.compare(metadata["metadata-version"], "1.1.0") < 0:
-            raise ValueError(
-                "this firmware image does not support setting the hub name"
-            )
-
-        name = name.encode() + b"\0"
-
-        max_size = metadata["max-hub-name-size"]
-
-        if len(name) > max_size:
-            raise ValueError(
-                f"name is too big - must be < {metadata['max-hub-name-size']} UTF-8 bytes"
-            )
-
-        offset = metadata["hub-name-offset"]
-        firmware[offset : offset + len(name)] = name
-
-    # Get checksum for this firmware
-    if metadata["checksum-type"] == "sum":
-        checksum = sum_complement(io.BytesIO(firmware), metadata["max-firmware-size"])
-    elif metadata["checksum-type"] == "crc32":
-        checksum = crc32_checksum(io.BytesIO(firmware), metadata["max-firmware-size"])
-    else:
-        raise ValueError(f"unsupported checksum type: {metadata['checksum-type']}")
-
-    # Append checksum to the firmware
-    firmware.extend(struct.pack("<I", checksum))
-
-    return firmware
-
-
-async def _create_firmware_v2(
-    metadata: AnyFirmwareV2Metadata, archive: zipfile.ZipFile, name: Optional[str]
-) -> bytearray:
-    base = archive.open("firmware-base.bin").read()
-
-    # start with base firmware binary blob
-    firmware = bytearray(base)
-
-    # Update hub name if given
-    if name:
-        name = name.encode() + b"\0"
-
-        max_size = metadata["hub-name-size"]
-
-        if len(name) > max_size:
-            raise ValueError(
-                f"name is too big - must be < {metadata['hub-name-size']} UTF-8 bytes"
-            )
-
-        offset = metadata["hub-name-offset"]
-        firmware[offset : offset + len(name)] = name
-
-    # Get checksum for this firmware
-    if metadata["checksum-type"] == "sum":
-        checksum = sum_complement(io.BytesIO(firmware), metadata["checksum-size"])
-    elif metadata["checksum-type"] == "crc32":
-        checksum = crc32_checksum(io.BytesIO(firmware), metadata["checksum-size"])
-    else:
-        raise ValueError(f"unsupported checksum type: {metadata['checksum-type']}")
-
-    # Append checksum to the firmware
-    firmware.extend(struct.pack("<I", checksum))
-
-    return firmware
-
-
-async def create_firmware(
-    firmware_zip: Union[str, os.PathLike, BinaryIO], name: Optional[str] = None
-) -> Tuple[bytes, AnyFirmwareMetadata]:
-    """Creates a firmware blob from base firmware and an optional custom name.
-
-    Arguments:
-        firmware_zip:
-            Path to the firmware zip file or a file-like object.
-        name:
-            A custom name for the hub.
-
-    Returns:
-        Tuple of composite binary blob for flashing and the metadata.
-
-    Raises:
-        ValueError:
-            A name is given but the firmware does not support it or the name
-            exceeds the alloted space in the firmware.
-
-    """
-
-    with zipfile.ZipFile(firmware_zip) as archive:
-        metadata: AnyFirmwareMetadata = json.load(
-            archive.open("firmware.metadata.json")
-        )
-
-        if firmware_metadata_is_v1(metadata):
-            firmware = await _create_firmware_v1(metadata, archive, name)
-        elif firmware_metadata_is_v2(metadata):
-            firmware = await _create_firmware_v2(metadata, archive, name)
-        else:
-            raise ValueError(
-                f"unsupported metadata version: {metadata['metadata-version']}"
-            )
-
-        return firmware, metadata
 
 
 # NAME, PAYLOAD_SIZE requirement